move all host ip related tests to test_host_ip_manager.py
