+++ managers/host_ip_manager.py
@@ -41,7 +41,8 @@ class HostIPManager:
     def store_host_ip(self) -> Optional[str]:
         """
         stores the host ip in the db
-        Retries to get the host IP online every 10s if not connected
+        recursively retries to get the host IP online every 10s if not
+        connected
         """
         if not self.main.db.is_running_non_stop():
             return
